Convert check programs in iocore/eventsystem/ into Catch based unit tests
(cherry picked from commit 39522703d7b9fdf4834b797e2d812f37f8970f71)

Conflicts:
	iocore/eventsystem/Makefile.am
	iocore/eventsystem/test_Buffer.cc
	iocore/eventsystem/test_Event.cc
